Javascript interface renamed following suggestion
